dev - Move abstract primary descriptor function to util package
