+++ src/bcfms/bcfms/pkg/extensions/functions/bc-fossil-publication-descriptors/bc_fossil_publication_descriptors.py
@@ -1,5 +1,5 @@
 from bcfms.util.bcfms_aliases import GraphSlugs, PublicationAliases as aliases
-from bcfms.functions.bc_primary_descriptors_function import BCPrimaryDescriptorsFunction
+from bcfms.util.bc_primary_descriptors_function import BCPrimaryDescriptorsFunction
 from bcfms.util.graph_lookup import GraphLookup
 
 details = {
