+++ src/common/util/bc_primary_descriptors_function.py
